+++ packages/contracts/contracts/TestContracts/SortedTrovesTester.sol
@@ -4,7 +4,6 @@ pragma solidity 0.6.11;
 
 import "../Interfaces/ISortedTroves.sol";
 
-
 contract SortedTrovesTester {
     ISortedTroves sortedTroves;
 
@@ -12,7 +11,12 @@ contract SortedTrovesTester {
         sortedTroves = ISortedTroves(_sortedTrovesAddress);
     }
 
-    function insert(address _id, uint256 _NICR, address _prevId, address _nextId) external {
+    function insert(
+        address _id,
+        uint256 _NICR,
+        address _prevId,
+        address _nextId
+    ) external {
         sortedTroves.insert(_id, _NICR, _prevId, _nextId);
     }
 
@@ -20,15 +24,20 @@ contract SortedTrovesTester {
         sortedTroves.remove(_id);
     }
 
-    function reInsert(address _id, uint256 _newNICR, address _prevId, address _nextId) external {
+    function reInsert(
+        address _id,
+        uint256 _newNICR,
+        address _prevId,
+        address _nextId
+    ) external {
         sortedTroves.reInsert(_id, _newNICR, _prevId, _nextId);
     }
 
-    function getNominalICR(address) external pure returns (uint) {
+    function getNominalICR(address) external pure returns (uint256) {
         return 1;
     }
 
-    function getCurrentICR(address, uint) external pure returns (uint) {
+    function getCurrentICR(address, uint256) external pure returns (uint256) {
         return 1;
     }
 }
